Implement log event parsing with optional count and filter parameters
Implements a new utility function for reading and parsing log files. Currently, the
implementation assumes that events in the log file will be separated by a newline or
carriage return. Also, assumes that log files will already be sorted chronologically.
Parsed logs are are returned as a list of Event objects, which currently only contain
a "raw" property for the entire event string.