Implement timestamp parsing and sorting of events
The timestamp parsing implementation depends on log events to be in Common Log Format
in order to extract the timestamp and sort events chronologically. If the timestamp
cannot be extracted, events are sorted as they were previously with events at the end
of the file treated as the newest. Also, the test log file was updated with events from
a server log in Common Log Format.